simple_image_analysis.py
#!/usr/bin/env python3
"""
Simple image analysis of FATMAP Mount Rainier image using PIL
To identify the climbing route line and extract coordinates
"""

from PIL import Image
import json

def analyze_image_colors():
    """Analyze the FATMAP image to understand colors and locate the route line"""
    
    try:
        # Load the image
        img = Image.open('static/images/mount_rainier.jpg')
        width, height = img.size
        print(f"Image dimensions: {width}x{height}")
        
        # Convert to RGB if needed
        if img.mode != 'RGB':
            img = img.convert('RGB')
        
        # Sample colors from different areas to understand the image
        sample_points = [
            # Bottom-left (Paradise area)
            (int(width * 0.1), int(height * 0.9)),
            (int(width * 0.2), int(height * 0.8)),
            
            # Center areas
            (int(width * 0.4), int(height * 0.6)),
            (int(width * 0.5), int(height * 0.5)),
            (int(width * 0.6), int(height * 0.4)),
            
            # Top-right (Summit area)
            (int(width * 0.8), int(height * 0.2)),
            (int(width * 0.9), int(height * 0.1)),
        ]
        
        print("\nSampling colors from key areas:")
        for i, (x, y) in enumerate(sample_points):
            r, g, b = img.getpixel((x, y))
            print(f"Point {i+1} ({x},{y}): RGB({r},{g},{b})")
        
        # Look for blue/green route colors in a more systematic way
        print("\nAnalyzing potential route colors...")
        
        # Define potential route line colors (blue/green/cyan variations)
        route_colors = []
        
        # Scan specific areas where routes might be
        scan_areas = [
            # Left side (Paradise to mid-mountain)
            {'x_range': (0.1, 0.4), 'y_range': (0.6, 0.9)},
            # Center (mid-mountain)
            {'x_range': (0.4, 0.7), 'y_range': (0.3, 0.7)},
            # Right side (upper mountain to summit)
            {'x_range': (0.7, 0.95), 'y_range': (0.05, 0.4)},
        ]
        
        unique_colors = set()
        for area in scan_areas:
            x_start = int(width * area['x_range'][0])
            x_end = int(width * area['x_range'][1])
            y_start = int(height * area['y_range'][0])
            y_end = int(height * area['y_range'][1])
            
            print(f"\nScanning area: x({x_start}-{x_end}), y({y_start}-{y_end})")
            
            for x in range(x_start, x_end, 10):  # Sample every 10 pixels
                for y in range(y_start, y_end, 10):
                    r, g, b = img.getpixel((x, y))
                    
                    # Look for bright/saturated colors that could be route lines
                    if (r + g + b) > 300 and (max(r, g, b) - min(r, g, b)) > 50:
                        unique_colors.add((r, g, b))
        
        # Sort colors by frequency/uniqueness
        color_list = list(unique_colors)
        print(f"\nFound {len(color_list)} unique bright colors:")
        for i, (r, g, b) in enumerate(color_list[:20]):  # Show first 20
            print(f"  Color {i+1}: RGB({r},{g},{b})")
        
        return img, width, height, color_list
        
    except Exception as e:
        print(f"Error analyzing image: {e}")
        return None, 0, 0, []

def find_route_manually():
    """Create manual route based on visual inspection"""
    
    print("\n" + "="*60)
    print("MANUAL ROUTE COORDINATE ANALYSIS")
    print("="*60)
    
    # Based on typical FATMAP climbing route visualization
    # Routes usually go from bottom-left to top-right in a curved path
    
    route_points = [
        # Starting from Paradise (typically bottom-left in mountain views)
        {"x": 10, "y": 90, "description": "Paradise Visitor Center"},
        {"x": 15, "y": 85, "description": "Paradise Valley"},
        {"x": 20, "y": 78, "description": "Panorama Point"},
        {"x": 28, "y": 70, "description": "Pebble Creek"},
        {"x": 35, "y": 62, "description": "Lower Muir Snowfield"},
        {"x": 42, "y": 54, "description": "Mid Muir Snowfield"},
        {"x": 48, "y": 46, "description": "Upper Muir Snowfield"},
        {"x": 54, "y": 38, "description": "Camp Muir"},
        {"x": 62, "y": 32, "description": "Cowlitz Glacier"},
        {"x": 70, "y": 28, "description": "Cathedral Gap"},
        {"x": 78, "y": 24, "description": "Disappointment Cleaver"},
        {"x": 85, "y": 18, "description": "Upper DC"},
        {"x": 90, "y": 12, "description": "Crater Rim"},
        {"x": 93, "y": 8, "description": "Columbia Crest Summit"},
    ]
    
    return route_points

def create_precise_coordinates():
    """Generate more precise coordinates for the climbing route"""
    
    # Load and analyze the image
    img, width, height, colors = analyze_image_colors()
    
    # Get manual route points
    route_points = find_route_manually()
    
    # Convert to the format needed by the app
    detailed_route = []
    elevations = [5400, 6000, 6800, 8200, 9000, 9500, 10000, 10188, 11000, 11800, 12800, 13200, 13800, 14411]
    
    zones = [
        "Paradise", "Paradise Valley", "Panorama Point", "Pebble Creek",
        "Lower Muir Snowfield", "Mid Muir Snowfield", "Upper Muir Snowfield", 
        "Camp Muir", "Cowlitz Glacier", "Cathedral Gap", "Disappointment Cleaver",
        "Upper DC", "Crater Rim", "Columbia Crest Summit"
    ]
    
    messages = [
        "🏁 Starting the ascent from Paradise!",
        "🌲 Climbing through Paradise meadows",
        "🌄 Beautiful views opening up!",
        "❄️ Entering the snowfield zone",
        "⛄ Steady climbing on snow",
        "⛄ Steep snow climbing ahead",
        "🏔️ Camp Muir coming into view",
        "⛺ Reached Camp Muir! Preparing gear...",
        "🧗 Technical glacier travel begins",
        "🪨 Approaching the rock band",
        "⚡ Most technical section - rock and ice!",
        "🎯 Pushing through the upper cleaver",
        "🌋 Almost to the crater rim!",
        "🏆 Summit achieved at 14,411 feet! 🎉"
    ]
    
    for i, point in enumerate(route_points):
        detailed_route.append({
            "x": point["x"],
            "y": point["y"],
            "zone": zones[i],
            "elevation": elevations[i],
            "message": messages[i]
        })
    
    return detailed_route

def main():
    """Main function to analyze the image and create route coordinates"""
    
    print("🏔️ FATMAP Image Analysis for Mount Rainier Route")
    print("="*60)
    
    # Create precise coordinates
    route = create_precise_coordinates()
    
    # Save to JSON
    with open('analyzed_route_coordinates.json', 'w') as f:
        json.dump({
            'method': 'PIL-based image analysis with manual refinement',
            'description': 'Mount Rainier Disappointment Cleaver route',
            'waypoints': route
        }, f, indent=2)
    
    print(f"\n✅ Generated {len(route)} route waypoints")
    print("📁 Saved to: analyzed_route_coordinates.json")
    
    # Generate JavaScript
    print(f"\n🚀 Updated JavaScript route for your app:")
    print("="*50)
    print("const routePath = [")
    for i, wp in enumerate(route):
        comma = "," if i < len(route) - 1 else ""
        print(f'    {{ x: {wp["x"]}, y: {wp["y"]}, zone: "{wp["zone"]}", elevation: {wp["elevation"]}, message: "{wp["message"]}" }}{comma}')
    print("];")

if __name__ == "__main__":
    main() 