requirements.txt
# Core Framework
gradio>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

# AI and RAG
openai>=1.3.0
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
sentence-transformers>=2.2.0
tiktoken>=0.5.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0

# Database
# sqlite3 is built into Python

# Environment Management
python-dotenv>=1.0.0

# API Integrations
httpx>=0.25.0
pydantic>=2.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
pillow>=10.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 